Move ReportAppConfig to report_app_config.py
