+++ corehq/apps/app_manager/models/__init__.py
@@ -105,7 +105,6 @@ from .modules import (  # noqa: F401
     ModuleBase,
     ModuleDetailsMixin,
     ParentSelect,
-    ReportAppConfig,
     ReportModule,
     ShadowFormEndpoint,
     ShadowModule,
@@ -120,6 +119,7 @@ from .report_app_config import (  # noqa: F401
     MobileFilterConfig,
     MobileSelectFilter,
     NumericFilter,
+    ReportAppConfig,
     ReportAppFilter,
     StaticChoiceFilter,
     StaticChoiceListFilter,
