+++ corehq/apps/app_manager/models/modules.py
@@ -8,11 +8,9 @@ from memoized import memoized
 
 from dimagi.ext.couchdbkit import (
     BooleanProperty,
-    DecimalProperty,
     DictProperty,
     DocumentSchema,
     IntegerProperty,
-    SchemaDictProperty,
     SchemaListProperty,
     SchemaProperty,
     StringProperty,
@@ -59,7 +57,6 @@ from .base import (
     rename_key,
 )
 from .case_list import (
-   GraphConfiguration,
    DetailColumn,
    Detail,
    CaseList,
@@ -82,7 +79,7 @@ from .mixins import (
     CommentMixin,
     NavMenuItemMediaMixin,
 )
-from .report_app_config import ReportAppFilter
+from .report_app_config import ReportAppConfig
 
 
 class ParentSelect(DocumentSchema):
@@ -811,43 +808,6 @@ class AdvancedModule(ModuleBase):
                 list(self.get_schedule_phases())[id].change_anchor(new_anchor)
             except IndexError:
                 pass  # That phase wasn't found, so we can't change it's anchor. Ignore it
-
-
-class ReportAppConfig(DocumentSchema):
-    """
-    Class for configuring how a user configurable report shows up in an app
-    """
-    # ID of the ReportConfiguration
-    report_id = StringProperty(required=True)
-    header = DictProperty()
-    localized_description = DictProperty()
-    xpath_description = StringProperty()
-    use_xpath_description = BooleanProperty(default=False)
-    show_data_table = BooleanProperty(default=True)
-    complete_graph_configs = DictProperty(GraphConfiguration)
-
-    filters = SchemaDictProperty(ReportAppFilter)
-    # Unique ID of this mobile report config
-    uuid = StringProperty(required=True)
-    report_slug = StringProperty(required=False)  # optional, user-provided
-    sync_delay = DecimalProperty(default=0.0)  # in hours
-
-    _report = None
-
-    def __init__(self, *args, **kwargs):
-        super(ReportAppConfig, self).__init__(*args, **kwargs)
-        if not self.uuid:
-            self.uuid = uuid.uuid4().hex
-
-    def report(self, domain):
-        if self._report is None:
-            from corehq.apps.userreports.models import get_report_config
-            self._report = get_report_config(self.report_id, domain)[0]
-        return self._report
-
-    @property
-    def instance_id(self):
-        return self.report_slug or self.uuid
 
 
 class ReportModule(ModuleBase):
